requirements.txt
django>=2.0
jsonfield
requests
cerberus
django-sendsms
